ranch feature/AIAI-389-Update-Agent-conversation-entry-with-guardrail-records - adapt OpenInferenceCollector to collect guardrail execution records
